refactor(tests): split test_config.py into a focused config package (#443)
tests/test_config.py was 1053 LOC and the largest remaining standalone
oversized test module after #896 (management_commands/), #912 (sync/),
#913 (utils/), #914 (models/), #917 (pr_command/), #918 (core_selectors/)
and #919 (cleanup/). It mixed nine separable concerns: overlay discovery,
load_config parsing, Mode resolution, per-overlay overrides,
check_for_updates, directory resolution, data/logging helpers and
e2e-repo config.

This splits it into a focused tests/config/ package, one file per
cohesive concern. The basename config is safe to use directly: it is a
test package, every import of teatree.config is absolute, and no
production code does a bare import config, so the package collects with
no shadowing (the full tests/ tree still collects 4272 tests).

Pure verbatim reorganization: every class and function is AST-identical
to the original (82 defs verified before and after ruff reordered
imports), only import lines changed and the two shared writers plus the
three shared fixtures were lifted verbatim into _shared.py / conftest.py.
Zero production code touched, zero assertions changed.

Relates-to #443
